feat!: Simplify `cmark(…)` by removing `State` parameter, introduce `cmark_resume(…)`.
The reason for doing it this way around causing a breaking change is
the change of behaviour for `cmark_resume`, which now returns a state
instance that needs a `finalize()` call to flush certain caches.

Currently this is only used to write link blocks, after which the state
is still usable for future invocations if necessary.

That way the caller has control over where to put link blocks, or other
items that should be placed at the end of a logical section.