Use necessary amount of backticks for inline code
We now insert the necessary number of backticks to quote any backticks
found in the code span itself. Before, only the case with a single
backtick was handled, now the code can contain arbitrarily many
backticks.

This was found using a fuzz test which tries to fuzz a single
paragraph of Markdown text. that is, it only fuzzes text which matches

    [Event::Start(Tag::Paragraph), .., Event::End(Tag::Paragraph)]