Fix emitting `::` for centered column in table.
Emitting `::` in a table to indicate alignment is `Center` seems to
break some parsers. Emitting `:-:` appears to work fine everywhere.

This commit also removes an allocation when encountering an empty header
cell. Rather than allocating a non-empty string, let the string be empty
and add a minimum width for emitting alignments.