refactor: Add helper to factor out newline + padding pattern
When writing a newline into the generated Markdown content
to start a new line of output, it is (almost) always necessary
to output the "padding" characters used to indent the content
at the current location in the document, based on the hierarchy
of block-level elements the output cursor is "inside" of.

Since writing a newline and writing the current padding are
always paired, factoring them out into a function should help
with readability and consistency in performing this minor
two-step dance correctly.