refactor: Add State::set_minimum_newlines_before_start()
This should help with readability. As an example, I didn't realize
initially that all of the `if` statements updated in this PR
were actually modifying the same field.

But once I started doing the more generic refactoring of adding
a `set_option_minimum()` method, I looked at every line and
realized they were all the same.

Factoring this to a method should clarify to future readers that
updating newlines_before_start specifically is a recurring operation.