feat: Improve CommonMark conformance testing output
This is a major improvement to the rendering of the failing
CommonMark conformance test output, that should make it easier
to read and understand when pulldown-cmark-to-cmark is producing
erroneous output.

Additionally, this introduces a way to force the printing of the
CommonMark failure report from the command line:

    $ FULL_CMARK_RESULTS=true cargo test

and documents in CONTRIBUTING.md that new contributors may want
to start with trying to improve conformance.

This removes `#[should_panic]` from the main CommonMark test. Instead,
we now hard-code in how many of the CommonMark tests are known to pass.

If the number that actually passes increases or decreases, this test
will fail, informing the user either that they have introduced a bug
(decrease), or that they've successfully improved conformance (increase).

Previously, if a change accidentally reduce the conformance, there was
no easy way to know, since that information was not stored/tracked
in any way.